Time: 58 ms (82.86%) | Memory: 56.2 MB (92.54%) - LeetSync

+++ 2581-divide-players-into-teams-of-equal-skill/divide-players-into-teams-of-equal-skill.cpp
@@ -10,7 +10,7 @@ public:
                 if(skill[i]+skill[n-i-1]!=sum){
                 return -1;
                 }
-                res+=(long long)skill[i]*skill[n-i-1];
+                res+=skill[i]*skill[n-i-1];
         }
         return res;
     }
